pyproject.toml
[project]
name = "taskflow-backend"
version = "0.1.0"
description = "TaskFlow Backend API (FastAPI + PostgreSQL)"
authors = [
    {name = "Mohammad Shebahati",email = "shebahati.mohammad@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
